requirements.txt
Django==1.9b1
Markdown==2.6.2
Pillow==2.9.0
mysqlclient==1.3.6
djangorestframework==3.3.0

